[FIX] 캐시 수정 및 추가

+++ src/main/java/com/sparta/quokkatravel/domain/common/cache/AccommodationCacheAspect.java
@@ -1,18 +1,19 @@
-package com.sparta.quokkatravel.domain.common.aop;
+package com.sparta.quokkatravel.domain.common.cache;
 
 import com.sparta.quokkatravel.domain.accommodation.repository.AccommodationRepository;
 import com.sparta.quokkatravel.domain.common.exception.NotFoundException;
 import com.sparta.quokkatravel.domain.room.entity.Room;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.annotation.AfterReturning;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
-import java.util.Set;
+import java.util.concurrent.TimeUnit;
 
 @Aspect
 @Slf4j
@@ -20,13 +21,15 @@ import java.util.Set;
 public class AccommodationCacheAspect extends AbstractCacheAspect {
 
     private final AccommodationRepository accommodationRepository;
+    private final RedissonClient redissonClient;
 
-    public AccommodationCacheAspect(RedisTemplate<String, Object> redisTemplate, AccommodationRepository accommodationRepository) {
+    public AccommodationCacheAspect(RedisTemplate<String, Object> redisTemplate, AccommodationRepository accommodationRepository, RedissonClient redissonClient) {
         super(redisTemplate);
         this.accommodationRepository = accommodationRepository;
+        this.redissonClient = redissonClient;
     }
     
-    @Before("@annotation(InvalidateAccommodationCache)")
+    @Before("@annotation(com.sparta.quokkatravel.domain.common.cache.InvalidateAccommodationCache)")
     public void evictAccommodationCache(JoinPoint joinPoint) {
         Long accommodationId = extractIdFromArgs(joinPoint);
         if (accommodationId == null) {
@@ -34,12 +37,23 @@ public class AccommodationCacheAspect extends AbstractCacheAspect {
             return;
         }
 
-        // Accommodation 캐시 삭제
-        evictCacheByPattern("Accommodation::" + accommodationId + "*");
+        RLock lock = redissonClient.getLock("accommodationLock: " + accommodationId);
+        try {
+            if (lock.tryLock(10, 3, TimeUnit.SECONDS)) {
+                // Accommodation 캐시 삭제
+                evictCacheByPattern("Accommodation::" + accommodationId + "*");
 
-        // 만약 삭제 작업일 경우에만 Room 캐시 삭제 수행
-        if (isDeleteOperation(joinPoint)) {
-            deleteAssociatedRoomCaches(accommodationId);
+                // 만약 삭제 작업일 경우에만 Room 캐시 삭제 수행
+                if (isDeleteOperation(joinPoint)) {
+                    deleteAssociatedRoomCaches(accommodationId);
+                }
+            } else {
+                log.warn("Failed to acquire lock for cache eviction on accommodation ID: {}", accommodationId);
+            }
+        } catch (InterruptedException e) {
+            log.error("Interrupted while trying to acquire lock for accommodation ID: {}", accommodationId, e);
+        } finally {
+            lock.unlock();
         }
     }
